refactor: move Paramore.Darker.QueryLogging sources into core (#321)
Tidy First (structural-only): relocate QueryLogging attributes and
decorators, Constants, and the Newtonsoft.Json-based builder extensions
from the side package `Paramore.Darker.QueryLogging` into core under
`src/Paramore.Darker/Logging/{Attributes,Handlers}/`. Type names keep the
`QueryLogging` prefix; only folders and namespaces rename to `Logging`.
The pre-existing `Logging/ApplicationLogging.cs` is unchanged. Side-package
csproj left empty pending Step 9 cleanup.
