+++ src/Paramore.Darker/Logging/Handlers/QueryLoggingDecoratorAsync.cs
@@ -4,11 +4,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using Paramore.Darker.Policies.Handlers;
 using Paramore.Darker.Exceptions;
-using Paramore.Darker.Logging;
+using Paramore.Darker.Policies.Handlers;
 
-namespace Paramore.Darker.QueryLogging
+namespace Paramore.Darker.Logging.Handlers
 {
     public class QueryLoggingDecoratorAsync<TQuery, TResult> : IQueryHandlerDecoratorAsync<TQuery, TResult>
         where TQuery : IQuery<TResult>
